fix permission

+++ packaging/makeself/build-static.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 # shellcheck source=./packaging/installer/functions.sh
@@ -13,15 +12,16 @@ case ${BUILDARCH} in
   x86_64) platform=linux/amd64 ;;
   armv7l) platform=linux/arm/v7 ;;
   aarch64) platform=linux/arm64/v8 ;;
+  ppc64le) platform=linux/ppc64le ;;
   *)
     echo "Unknown target architecture '${BUILDARCH}'."
     exit 1
     ;;
 esac
 
-DOCKER_CONTAINER_NAME="netdata-package-${BUILDARCH}-static-alpine314"
+DOCKER_CONTAINER_NAME="netdata-package-${BUILDARCH}-static-alpine315"
 
-if [ "${BUILDARCH}" != "$(uname -m)" ]; then
+if [ "${BUILDARCH}" != "$(uname -m)" ] && [ "$(uname -m)" = 'x86_64' ] && [ -z "${SKIP_EMULATION}" ]; then
     docker run --rm --privileged multiarch/qemu-user-static --reset -p yes || exit 1
 fi
 
@@ -39,12 +39,12 @@ if ! docker inspect "${DOCKER_CONTAINER_NAME}" > /dev/null 2>&1; then
   # inside the container and runs the script install-alpine-packages.sh
   # (also inside the container)
   #
-  if docker inspect alpine:3.14 > /dev/null 2>&1; then
-    run docker image remove alpine:3.14
-    run docker pull --platform=${platform}  alpine:3.14
+  if docker inspect alpine:3.15 > /dev/null 2>&1; then
+    run docker image remove alpine:3.15
+    run docker pull --platform=${platform}  alpine:3.15
   fi
 
-  run docker run --platform=${platform} -v "$(pwd)":/usr/src/netdata.git:rw alpine:3.14 \
+  run docker run --platform=${platform} -v "$(pwd)":/usr/src/netdata.git:rw alpine:3.15 \
     /bin/sh /usr/src/netdata.git/packaging/makeself/install-alpine-packages.sh
 
   # save the changes made permanently
